Don't export ECMetadataProvider outside presentation-components
