+++ presentation/components/src/presentation-components.ts
@@ -113,7 +113,6 @@ export * from "./presentation-components/instance-filter-builder/InstanceFilterB
 export * from "./presentation-components/instance-filter-builder/PresentationInstanceFilterBuilder";
 export * from "./presentation-components/instance-filter-builder/PresentationInstanceFilterProperty";
 export * from "./presentation-components/instance-filter-builder/InstanceFilterConverter";
-export * from "./presentation-components/instance-filter-builder/ECMetadataProvider";
 export * from "./presentation-components/instance-filter-builder/Types";
 export * from "./presentation-components/instance-filter-builder/Utils";
